Moves middleware folder inside rest folder
