+++ src/shared/libs/rest/middleware/private-route.middleware.ts
@@ -1,7 +1,7 @@
 import { StatusCodes } from 'http-status-codes';
 import { NextFunction, Request, Response } from 'express';
 import { Middleware } from './middleware.interface.js';
-import { HttpError } from '../libs/rest/index.js';
+import { HttpError } from '../index.js';
 
 export class PrivateRouteMiddleware implements Middleware {
   public async execute({ tokenPayload }: Request, _res: Response, next: NextFunction): Promise<void> {
