+++ src/shared/libs/rest/middleware/types/document-exists.interface.ts
