+++ src/shared/models/rent/rent.controller.ts
@@ -8,10 +8,10 @@ import { fillDTO, makeCapitalized } from '../../helpers/common.js';
 import { RentRdo } from './rdo/rent.rdo.js';
 import { ParamsRentId } from './index.js';
 import { QueryRent } from './type/query-rent.type.js';
-import { ValidateObjectIdMiddleware } from '../../middleware/validate-objectid.middleware.js';
-import { ValidateDtoMiddleware } from '../../middleware/validate-dto.middleware.js';
-import { DocumentExistsMidleware } from '../../middleware/document-exists.middleware.js';
-import { PrivateRouteMiddleware } from '../../middleware/private-route.middleware.js';
+import { ValidateObjectIdMiddleware } from '../../libs/rest/middleware/validate-objectid.middleware.js';
+import { ValidateDtoMiddleware } from '../../libs/rest/middleware/validate-dto.middleware.js';
+import { DocumentExistsMidleware } from '../../libs/rest/middleware/document-exists.middleware.js';
+import { PrivateRouteMiddleware } from '../../libs/rest/middleware/private-route.middleware.js';
 import { UserService } from '../user/user-service.interface.js';
 import { StatusCodes } from 'http-status-codes';
 import { DocumentType } from '@typegoose/typegoose';
