+++ src/shared/models/user/user.controller.ts
@@ -9,13 +9,13 @@ import { UserRdo } from './rdo/user.rdo.js';
 import { StatusCodes } from 'http-status-codes';
 import { Config, RestSchema } from '../../libs/config/index.js';
 import { ParamsUserId } from './index.js';
-import { ValidateObjectIdMiddleware } from '../../middleware/validate-objectid.middleware.js';
-import { ValidateDtoMiddleware } from '../../middleware/validate-dto.middleware.js';
-import { DocumentExistsMidleware } from '../../middleware/document-exists.middleware.js';
-import { UploadFileMiddleware } from '../../middleware/upload-file.middleware.js';
+import { ValidateObjectIdMiddleware } from '../../libs/rest/middleware/validate-objectid.middleware.js';
+import { ValidateDtoMiddleware } from '../../libs/rest/middleware/validate-dto.middleware.js';
+import { DocumentExistsMidleware } from '../../libs/rest/middleware/document-exists.middleware.js';
+import { UploadFileMiddleware } from '../../libs/rest/middleware/upload-file.middleware.js';
 import { AuthService } from '../auth/index.js';
 import { LoginRdo } from './rdo/login.rdo.js';
-import { PrivateRouteMiddleware } from '../../middleware/index.js';
+import { PrivateRouteMiddleware } from '../../libs/rest/middleware/index.js';
 
 @injectable()
 export class UserController extends AbstractController {
